requirements.txt
chromadb==1.0.13
fastapi==0.121.2
PyPDF2==3.0.1
python-docx==1.2.0
python-multipart==0.0.20
sentence-transformers==5.1.2
uvicorn==0.34.3